Move StateTable to runners/core-java
